Start towards 0.9.0. Change to using PalmOS SD 3.5 and prc-tools-2.0 to build: this changes a lot of types through the code, and also the way we build. Don't store resources in a subdirectory while building, because this seems to break the current build-prc. Start adding support for export to the MemoPad.

+++ keyring/passwd.h
@@ -27,6 +27,6 @@ void Unlock_ObliterateKey(void);
 Boolean UnlockForm_HandleEvent(EventPtr event);
 Boolean UnlockForm_Run(void);
 Boolean Unlock_CheckTimeout(void);
-void Unlock_SetExpiry(ULong newExpiry);
+void Unlock_SetExpiry(UInt32 newExpiry);
 void Unlock_Reset(void);
 void Unlock_PrimeTimer(void);
